requirements.txt
requests
beautifulsoup4
ruamel.yaml
pydantic
